Add smtp_sasl_password_maps to test for env_var
Remove packages that are not in focus of a streamline box

Install PHPUnit via Curl, on system build

Add HTML_QuickForm for web app

+++ run.sh
@@ -40,26 +40,33 @@ mv /opt/ssl-cert-snakeoil.key /data/apache2/ssl/ssl-cert-snakeoil.key
 mv /opt/ssl-cert-snakeoil.pem /data/apache2/ssl/ssl-cert-snakeoil.pem
 
 
-# Tweak Apache build
-sed -i 's|\[PHP\]|\[PHP\] \nIS_LIVE=1 \nIS_DEV=1 \n;The IS_DEV is set for testing outside of DEV environments ie: test.domain.tld|g' /etc/php5/apache2/php.ini
-sed -i 's|;include_path = ".:/usr/share/php"|include_path = ".:/usr/share/php:/data/pear"|g' /etc/php5/apache2/php.ini
-sed -i "s/variables_order.*/variables_order = \"EGPCS\"/g" /etc/php5/apache2/php.ini
+if [ ! -f /etc/php5/apache2/build ]; then
 
-# Update the PHP.ini file, enable <? ?> tags and quieten logging.
-sed -i "s/short_open_tag = Off/short_open_tag = On/" /etc/php5/apache2/php.ini
-sed -i "s/error_reporting = .*$/error_reporting = E_ERROR | E_WARNING | E_PARSE/" /etc/php5/apache2/php.ini
-sed -i 's|;session.save_path = "/var/lib/php5"|session.save_path = "/tmp"|g' /etc/php5/apache2/php.ini
-sed -i 's|"\/var\/log\/apache2\/error.log"|"\/data\/apache2\/error.log"|g' /etc/php5/apache2/php.ini
+    # Tweak Apache build
+    sed -i 's|\[PHP\]|\[PHP\] \nIS_LIVE=1 \nIS_DEV=1 \n;The IS_DEV is set for testing outside of DEV environments ie: test.domain.tld|g' /etc/php5/apache2/php.ini
+    sed -i 's|;include_path = ".:/usr/share/php"|include_path = ".:/usr/share/php:/data/pear"|g' /etc/php5/apache2/php.ini
+    sed -i "s/variables_order.*/variables_order = \"EGPCS\"/g" /etc/php5/apache2/php.ini
 
-sed -i 's|#ServerRoot "\/etc\/apache2"|ServerRoot "\/data\/apache2"|g' /etc/apache2/apache2.conf
+    # Update the PHP.ini file, enable <? ?> tags and quiet logging.
+    sed -i "s/short_open_tag = Off/short_open_tag = On/" /etc/php5/apache2/php.ini
+    sed -i "s/error_reporting = .*$/error_reporting = E_ERROR | E_WARNING | E_PARSE/" /etc/php5/apache2/php.ini
+    sed -i 's|;session.save_path = "/var/lib/php5"|session.save_path = "/tmp"|g' /etc/php5/apache2/php.ini
+    sed -i 's|"\/var\/log\/apache2\/error.log"|"\/data\/apache2\/error.log"|g' /etc/php5/apache2/php.ini
 
-# Allow the container to continuously update it's time
-echo "ntpdate ntp.ubuntu.com" > /etc/cron.daily/ntpdate && chmod 755 /etc/cron.daily/ntpdate
+    sed -i 's|#ServerRoot "\/etc\/apache2"|ServerRoot "\/data\/apache2"|g' /etc/apache2/apache2.conf
 
+    # Allow the container to continuously update it's time
+    echo "ntpdate ntp.ubuntu.com" > /etc/cron.daily/ntpdate && chmod 755 /etc/cron.daily/ntpdate
+
+    # Add build file to remove duplicate script execution
+    echo 1 > /etc/php5/apache2/build
+
+fi
 
 # Postfix uses smart hosts in cluster to relay email
 postconf -e \
 	relayhost=[post-office.htmlgraphic.com]:25 \
+    smtp_sasl_password_maps=static:$USER:$PASS \
 	inet_protocols=ipv4
 
 # Postfix is not using /etc/resolv.conf is because it is running inside a chroot jail, needs its own copy.
